Fix reverse proxy duplicate requestID result wrong response.
When as a proxy server, this program will reuse the requestID from client,
and the requestID is the communication identifier for a tcp channel.
The reverse proxy server will have the same requestID from different clients,
but those requests may use a same channel when send to then backend server.
Here we use a different id as the communication identifier.